Finished applying textures, tried to fix chair
All the materials I’ve made are in. The chair has some issues
with its normals - from behind, it appears to have no back.
I made a copy of the chair, modified it in maya to fix the problem, and
reimported it, but the texture info was lost. Anyone know how to edit a
mesh in maya without compromising the texture?